backend/requirements.txt
pydantic-core
openai
requests_toolbelt
python-docx